Fixed bugs (#69)

+++ src/main/java/md2k/mcerebrum/CSVParser.java
@@ -44,7 +44,7 @@ public class CSVParser implements Iterable<CSVDataPoint> {
         CSVDataPoint tempPacket;
 
         String[] tokens;
-        int data;
+        double data;
         long timestamp;
 
         File file = new File(filename);
@@ -55,7 +55,7 @@ public class CSVParser implements Iterable<CSVDataPoint> {
                 tokens = scanner.nextLine().split(",");
                 double ts = Double.parseDouble(tokens[0]);
                 timestamp = (long) ts;
-                data = (int) Double.parseDouble(tokens[1]);
+                data = Double.parseDouble(tokens[1]);
 
                 tempPacket = new CSVDataPoint(channel, timestamp, data);
                 this.data.add(tempPacket);
